User ids that begin with 0 cannot authenticate through ldap
Currently, in the ldap2py function, several fields are attempted
to be converted to python friendly types.

In doing so, an attempt to convert a field to int() is attempted,
but in some cases, a user id may begin with a 0. When the user
attempts to authenticate, they will be rejected since any additional
query will use the id without the 0 in front.

Closes-Bug: #1396763

Change-Id: I1e2436b845e534f6cdb0398b5cca17d8502b905f
(cherry picked from commit 474271683f5e44c6253b75102026cc8578c10d06)

Conflicts:
	keystone/tests/unit/common/test_ldap.py